[DO NOT MERGE] Proof of concept code for ipam
This code is not expected to be merged as is, because it is only proof
of concept at this point.
However parts of this code are planned to be used in future production
version of Infoblox IPAM Driver.
Uploading code into gerrit to share with other team members.

<2015/08/07>
Refactor InfobloxObjectManager to eliminate the IPv4/6 backends. Instead,
create InfobloxObjects to encapsulate the IPv4/6 differences.
NOTE: this is just a partial commit to get feedback.

Change-Id: Ib106e9d2133794039ff05427417c4a396d48cb0b