Add IP allocator module
Use common interface for allocating ip addresses for HostRecords and
FixedAddress.
Includes logic for binding/unbinding names with ip allocation.

Decidion to use host record or not has to be done outside IP Allocator.
So for each ip allocation new instance of IP Allocator has to be
created to be able to change allocation class from allocation to
allocation. It is needed to address [1].

[1] https://twiki.inca.infoblox.com/do/view/Main/
IOAIPAMDesignLibertyHighLevel#2_8_IP_Aliasing_Support_for_Mult

Change-Id: I6270bd9e94f363de3664357566a897e8bf0f5137