Pattern Builder
Pattern Builder is used to create a hostname or zone name for DNS.

A hostname can be created by ip address or instance name with a prefix
indicating such a device owner type as 'floatingip', 'dhcp-port',
'router-iface', and so on.

For example, if the device owner is 'dhcp-port' and the default hostname
configuration is set to use '{ip-address}, a final hostname would be something
like 'dhcp-11.11.1.1'.

A zone name uses the default domain suffix configuration which can be
statically defined or dynamically defined with subnet or network name.

Any dynamic property is defined with brackets like {subnet_name} and
{network_name}.

For example, the zone name could be '{subnet_name}.infoblox.com' where
'{subnet_name}' portion will be replaced with a real value.

The fully qualified domain name would be the combination of hostname and
zone name. For example, 'dhcp-11.11.1.1.subnet-001.infoblox.com'

Change-Id: I2edd1101d7a3f69e4ba5e6450cc7d9eb353aae24