[DO NOT MERGE] Update allocation pools in driver
Reference IPAM driver updates subnet pools by deleting all allocation
pools for subnet and creating new pools.
But in case of NIOS deleting all ip_ranges deletes all ip
allocations for subnet, so exploring and testing how harmfull update
subnet might become in normal usage.

Several options I see are:
- calculate difference between new ranges and old ranges and apply
  changes.
- do not allow allocation pool update after subnet is created.

This patch is work-in-progress version of code and not expected to be
merged as is.

Change-Id: If77aa542f1fb36b943c3d925efc84e7c2db9c7ef