DNS Controller
DNS controller creates/deletes dns zones and binds/unbinds FQDN to
and from an ip address to create/delete DNS records.

Domain suffix name is used to generate a FQDN and it can be configured
to use either static name or a dynamic name using various patterns.
A static name is like "infoblox.com" as a literal name.
A dynamic name is a pattern that can be replaced with a real value.
The allowed dynamic patterns are {subnet_id}, {subnet_name}, {network_id},
{network_name}, {tenant_id}, {tenant_name}, {address_scope_id},
and {address_scope_name}.

When a neutron object name is used rather than object id like
{subnet_name}, {network_name}, and so on, it is the user's responsibilty
that neutron object names have to be unique.

When deleting a dns zone, the controller checks whether the zone can be
removed or not by looking at any dependency. For example, deleting a subnet
action can trigger a deleting a zone. In this case, we check what pattern is
used for zone name. If a network id or name is used, the zone will not be
removed when more than one subnets exist. The zone will be removed when
only one subnet remains so that no subnet is sharing the zone name.

Change-Id: I1f8df0cbc480e6832a380da52f07a03f445ac7c6