Fix DCHP Port IP Generation Issue
When DHCP Port IP is generated, it should create a host record only with dhcp
disabled option but creates DNS records as well when IP Allocation Strategy is
set to 'Fixed Address' with DNS binging options.

Added dhcp port check on bind and unbind names method in dns controller so
that it can ignore dns operation for dhcp port ip.

Change-Id: Ib3b440d959de059bb18e01fc96f34f0c216ef577