Fix 'IP Type' EA for Floating IP and Fixed IP
When an instance is getting an ip, the device owner is 'compute:nova'.
The ip may be a fixed ip from a private subnet or a floating ip from the
external network. Only notification handler can tell whether ip is for
floating ip association or not. Thus 'is_floating_ip' parameter is introduced
and set to True only for the floating ip association case.

Fixed https://jira.inca.infoblox.com/browse/OPENSTACK-745

Change-Id: I7b220e0e54c3ab583c401ece231f4e9c18de38e1