Populate grid_primary on zone creation
Previously grid_primary was not populated and it cause
failure on zone creation.
For delegated zones NIOS requires CPM member to be listed in grid
primaries.
Current approach is to build grid_primary from authority_member.

But looks like authority member from context always points to GM,
so this patch is only partial fix for the issue.

Partialy fixes OS-746 [1].

1. https://jira.inca.infoblox.com/browse/OPENSTACK-746

Change-Id: I1af047b07acfb5b28842aad6e454cd65cc503171