Do not use network_exists
*WIP*
network_exists is quite ineffective way of using check network existence
information.
It makes search wapi call, but not return network if it was found.
So if network exists and some update action has to be performed on it,
then it should be searched once again.
This is the reason why it was deprecated in infoblox-client.

Reworked code to use objects directly and reduce count of api calls to
NIOS.

TODO: Need to rework UT to comply with new workflow.

Change-Id: Ifbecbe96669987c51ffd4da8c16f4bbfa43fcae0