Collect tenant id and tenant name info
Tenant name is not part of neutron and belongs to keystone,
so this information can not be queried from neutron.

To build mapping between tenant ids and tenant names this patch
listens to notification events related to network creation and
saves explored mapping to database.

There are cases where network tenant_id mismatch with context tenant_id.
To resolve this case we check existent mapping in database.
If network tenant id is not there, then last resort action is called:
API call to keystone is done with auth_token obtained from context.

Change-Id: I46dfa73ca246e2b64510a246403513fa5bace82b