Add "VM Name" EA for fip associated with instance
Association of floating ip is handled in update_floatinpip_sync()
notification hander. Unfortunately, the payload does not contain
the name of the instance.
The payload does, however, include the fixed ip address of the
associated instance. The code therefore uses this to query Infoblox
for the object associated with the instance. The instance name
is then obtained from the resulting object.

Change-Id: I5e108d27c083824d7edb79c3e6591624979bf22f