Cloud Member Regisration
"Is Cloud Member" EA is introduced for the grid registration.
The EA should be set to True for each member in the grid if a member
should be used in OpenStack. The grid master member is an exception.
Whether the EA is set or not, the grid master will be registered.

Change-Id: Idd2b32cf0f4f2c73c19485c34a9f19dcb5ae5cb0