Fix Port Device Owner for Instance Creation
When an instance is created from CLI using nova-boot, device owner
field is not populated correctly. It should be 'compute:nova' but
'compute:None' is passed.

The issue is coming from nova community code but we need to take care of
this in our agent to update port related EAs correctly.

Change-Id: Ie721c25fb2aae770c5f67c43abaaeefe7fd89f5e