Host Record DNS Record Binding Not Needed in Instance Name Sync
When an instance is created, a Host record is created.
Then notification handler subscribes the instance creation message and
tries to update the host record with instance info such as 'VM Name'
and 'VM ID'.

The host record ip allocator did not check if the record exists or not
correctly. The object's 'ref' value should be used.

This incorrect check ends up executing wrong logic and removes the host
record.

The 'ref' value check fixes this issue.

Also when this error happens, the instance cannot be removed
from Openstack side.
The driver now does not throw the exception when the host record is not
found and let the neutron to continue its process.

Change-Id: Ib0314e9cda0ef8246c4dbe660f5433205d55d419