Host Record DNS Record Binding Not Needed in Instance Name Sync
When an instance is created, a Host record is created.
Then notification handler subscribes the instance creation message and
tries to bind the host record with instance name.
However, the {instance_name} pattern is not used.
This results in an error in HostRecordIPAllocator and ends up removing the
host record.

If {instance_name} pattern is not used, no need to bind the host record
with instance name in the notification handler.

Also when this happens, the instance cannot be removed from Openstack side.
The driver now does not throw the exception when the host record is not
found and let the neutron to continue its process.

Change-Id: Ib0314e9cda0ef8246c4dbe660f5433205d55d419