Fix External/Shared Subnet Deletion
When an external or a shared subnet is deleted, cloud EAs that contains subnet
information on infoblox network, range, and zones need to be reset.

The mandary EAs cannot be empty. So 'CMP Type' and 'Tenant ID' will be set to
'N/A' and 'Cloud API Owned' will be set to 'False'.

Also the network mapping entry will be removed since subnet no longer exists.

Change-Id: Ib5aee6606d2b75694ec5be08f4f88ede2d67213a