Support IPAM Only
In order to support pure IPAM only, we should not create any zones.
Also when we create a host record for DHCP port ip, we should set
'configure_for_dns' option to False so that the host record can be
created without dns zones.

Please note that this change requires a fix from infoblox-client:
https://github.com/infobloxopen/infoblox-client/issues/94

Change-Id: Ib09ca29c78c08a026f35a17b8a1a3913a79ace5b