Max Result Set Configuration Support
The ipam driver threw "Result set too large" when retrieving records exceeds
1000 since that is the default NIOS setting.

'wapi_max_results' configuration parameter is now added for the user to
control this value in neutron.conf.

According to WAPI document, it can accept a negative or positive number.

"Mximum number of objects to be returned. If set to a negative number the
appliance will return an error when the number of returned objects would
exceed the setting. The default is -1000. If this is set to a positive
number, the results will be truncated when necessary."

Change-Id: I103d654e826b8b8854d5fd99e6032718209b0079