Network View Selection for OpenStack and Default Network View
To allow the Infoblox grid to use for multiple adapters, we introduce
"Cloud Adapter ID" EA that can be set on a network view by the user.

The cloud adater id is the value used in "cloud_data_center_id" configuration
parameter in neutron.conf.

The user needs to explicitly set this EA on each network view that
is intended to use in OpenStack.

Also a default network view may not always be 'default' since the user
can change the name or create a different default network view.

During the ipam agent sync, the default network view is detected and
stored in the database as a part of the network view definition.

Change-Id: I5e5e5d174d0b0e3c3b532f78b38171208016127c