Network View ID Change and GM Member Registration
Network View ID currently uses a db generated ID and the ID is stored in
the Infoblox grid to use for network view sync by the ipam agent.

This requires the WRITE permission on network views and this becomes an
issue in some organization. In order to remove this limitation, now
we use the NIOS object ID of a Network View. This does not require
a WRITE permission on network views.

Also the GM member usability has been enhanced.

'grid_master_host' in neutron.conf is to provide GM member WAPI connectivity.
Before this change, we enforced this information to match what registered
in the Infoblox grid but this restriction is removed.

GM member can be replaced since a GM Candidate (GMC) can be promoted to
be the GM. When this happens, the GM member ID would not change to make
GMC promotion a smooth operation.

Change-Id: Ie27f0149e8d54357e77102c575c7f721965dfcda