Fix Network View ID Collision
Network View ID currently uses NIOS object ID and this is a primary key in
infoblox_network_views table.

As it turns out, NIOS object ID for 'default' network view is always the same
even in a different grid.

When another grid is registered in OpenStack, the new grid sync fails because
the 'default' network view id is not unique.

In order to prevent this issue, now the network view id in
infoblox_network_views table is generated with the combination of grid id and
NIOS object ID.

Change-Id: I69fedc0caf6f2718a1253970a3799e0e8a4b472b