Data Migratoin Support for IPAM Driver
OpenStack could have been used without the Infoblox ipam driver.
Later, when the user decides to install networking-infoblox and uses it as
the neutron ipam driver, the existing neutron objects such as networks,
subnets and ports need to be migrated to the infoblox grid.

This script will read neutron objects and convert them to Infoblox objects
in the Infoblox grid.

Change-Id: I44decbeb0f98409d66f1df1ab2281f8540d73c41