Fix returning default fields for Member
Passing '[]' as return fields for Member causes returning default fields
for current wapi version. For newer wapi version (2.2 and 2.3) default
fields for Member were expanded and not all of them are currently
supported by infoblox-client, so it caused failure on parsing reply.
Passing 'host_name' as return_fields to get single field in reply
instead of list of default fields.

Change-Id: I3665aa0f2d861606f0dbd070a2edd21b7e14a706