Allow use of port_name in host name pattern
When specifying a host name pattern, it would be very useful to be able
to use the port_name in the pattern. That would enable differentiation
of DNS names for different ports on the same instance, without resorting
to the IP octets.
Added code to support 'port_name' in host name pattern.

Change-Id: I6424ba23e34884689637d410ae10b1f652971bb0