Use mock to prevent plugin initialization
In this test case there is no need to initialize plugin,
but if None is passed to IpamEventHandler it gets initialized.

Observed that in some environmets it leads to test failures.
So use mock instead of plugin in this test.

Also remove _setup_config since plugin is no longer loaded and do not
require configuration.

Change-Id: Ia0bdbc9f7c7ba903802f2a20f2168d969e193e59