Allow range modification for shared network
Previously, the driver would refuse to deleting any IP range from Infoblox if
the containing network is shared/external.

This causes range modification
to fail if, for example, the user tries to extend an existing range - since
this would require deleting the existing range and creating a new one.

To work around this, the code checks if the IP range we are trying to
delete include an EA that indicates that the range is created by the
IPAM driver. If so, the driver will delete IP range just like the
case for private network.
If the EA does not exist, the IPAM driver assumes that the range is
created manually outside of OpenStack - in which case, the IP range
will not be deleted.

Change-Id: I7b8a69651c00c12d6dd3a3d177a21984279ce586