Delete Fixed IP if neutron port unknown
Added --delete-unknown-ips command line option.

If the option is specified, Fixed IPs in Infoblox that does not
correspond to a known neutron port will be deleted. NOTE that this
only applies to Fixed IP in private network. IOW, any Infoblox
network that has 'Is Shared' or 'Is External' EA set to true will
not be processed.

Change-Id: I3309703d03bf0320042166067ed3fd8234ec952a