Fix issue in the rollback mechanism causes DNS zone delete in NIOS
Added check to ignore rollback of objects which are pre-exist.
Raised pull request in infoblox client:
https://github.com/infobloxopen/infoblox-client/pull/162
which support create_check_exist function to return object and
obj_created flag to check if InfobloxObj is created or reused.

Updated requirement.txt to get latest version of infoblox_client
for which pull request is raised:
https://github.com/infobloxopen/infoblox-client/pull/163

Change-Id: I54bed6d64a4b03be8d238ac9556a89f65270facd
Closes-Bug: #1688347