Handled exception in ip allocation for IPV6.
Fixed issue when dns host record for a dhcp ip is avaiable for the port
and ip_allocation request for the ip comes.
This case occurs if we delete only network from the NIOS and try to
sync the network using sync tool.

This fix requires changes in infoblox_client.
Raised PR at: https://github.com/infobloxopen/infoblox-client/pull/166

Change-Id: I9ef0d79522a7423ea2093a6253e4c6e491d98482