Update tox_install.sh to align for sphinx jobs
The updates to the sphinx docs jobs in support of the updates to
the PTI wound up exposing an unintended interface. There are two flavors
of the tox_install.sh file out there, and we basically need to collapse
them into one flavor.

Update the tox_install.sh script to match the
constraints-as-first-argument form.

Update tox.ini to ignore temporary .tmp directory.

Change-Id: I92e7250b1375d4995683dae0347ebb0843bf9b59